# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maplog - Soundmap companion archive API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Song {
  id: number;
  title: string;
  artist: string;
  /** @nullable */
  album?: string | null;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  appleMusicId?: string | null;
  cardCount?: number;
  createdAt: string;
}

/**
 * JSON visual config: borderColor, glowColor, backgroundEffect, badgeLabel, hasShimmer, hasParticles, etc.
 */
export type RarityTypeThemeConfig = { [key: string]: unknown };

export interface RarityType {
  id: number;
  slug: string;
  name: string;
  category: string;
  /** Numeric tier for sorting (higher = rarer) */
  tier: number;
  /** JSON visual config: borderColor, glowColor, backgroundEffect, badgeLabel, hasShimmer, hasParticles, etc. */
  themeConfig: RarityTypeThemeConfig;
}

export interface CollectedCard {
  id: number;
  songId: number;
  rarityTypeId: number;
  rarityType: RarityType;
  /**
     * e.g. '#042', 'Halloween', 'Freshman'
     * @nullable
     */
  variantLabel?: string | null;
  /** @nullable */
  artworkUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  collectedAt: string;
}

export interface SongDetail {
  id: number;
  title: string;
  artist: string;
  /** @nullable */
  album?: string | null;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  appleMusicId?: string | null;
  createdAt: string;
  cards: CollectedCard[];
}

export interface SongInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  artist: string;
  album?: string;
  durationMs?: number;
  audioUrl?: string;
  appleMusicId?: string;
}

export interface SongUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  artist?: string;
  album?: string;
  durationMs?: number;
  audioUrl?: string;
  appleMusicId?: string;
}

export type RarityTypeInputThemeConfig = { [key: string]: unknown };

export interface RarityTypeInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  tier: number;
  themeConfig: RarityTypeInputThemeConfig;
}

export type RarityTypeUpdateThemeConfig = { [key: string]: unknown };

export interface RarityTypeUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  tier?: number;
  themeConfig?: RarityTypeUpdateThemeConfig;
}

export interface CollectedCardInput {
  songId: number;
  rarityTypeId: number;
  variantLabel?: string;
  artworkUrl?: string;
  notes?: string;
}

export interface CollectedCardUpdate {
  variantLabel?: string;
  artworkUrl?: string;
  notes?: string;
}

export interface Playlist {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  songCount: number;
  createdAt: string;
}

export interface PlaylistDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  songCount: number;
  createdAt: string;
  songs: Song[];
}

export interface PlaylistInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface PlaylistUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export interface PlaylistSongInput {
  songId: number;
}

export interface CategoryStat {
  category: string;
  count: number;
  tier: number;
}

export interface CollectionStats {
  totalSongs: number;
  totalCards: number;
  totalPlaylists?: number;
  byCategory: CategoryStat[];
  recentlyAdded: Song[];
}

export type ListSongsParams = {
search?: string;
artist?: string;
/**
 * Filter to songs that have at least one collected card of this rarity category
 */
rarityCategory?: string;
limit?: number;
offset?: number;
};

export type ListRarityTypesParams = {
category?: string;
};

export type ListCollectedCardsParams = {
songId?: number;
rarityTypeId?: number;
category?: string;
limit?: number;
offset?: number;
};