# lib/api-zod/src/generated/types/collectedCard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maplog - Soundmap companion archive API
 * OpenAPI spec version: 0.1.0
 */
import type { RarityType } from './rarityType';

export interface CollectedCard {
  id: number;
  songId: number;
  rarityTypeId: number;
  rarityType: RarityType;
  /**
     * e.g. '#042', 'Halloween', 'Freshman'
     * @nullable
     */
  variantLabel?: string | null;
  /** @nullable */
  artworkUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  collectedAt: Date;
}