# lib/api-zod/src/generated/types/collectedCardInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maplog - Soundmap companion archive API
 * OpenAPI spec version: 0.1.0
 */

export interface CollectedCardInput {
  songId: number;
  rarityTypeId: number;
  variantLabel?: string;
  artworkUrl?: string;
  notes?: string;
}