# lib/api-zod/src/generated/types/playlistDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maplog - Soundmap companion archive API
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface PlaylistDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  songCount: number;
  createdAt: Date;
  songs: Song[];
}