# lib/api-zod/src/generated/types/playlistInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maplog - Soundmap companion archive API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}