# lib/api-zod/src/generated/types/rarityType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maplog - Soundmap companion archive API
 * OpenAPI spec version: 0.1.0
 */
import type { RarityTypeThemeConfig } from './rarityTypeThemeConfig';

export interface RarityType {
  id: number;
  slug: string;
  name: string;
  category: string;
  /** Numeric tier for sorting (higher = rarer) */
  tier: number;
  /** JSON visual config: borderColor, glowColor, backgroundEffect, badgeLabel, hasShimmer, hasParticles, etc. */
  themeConfig: RarityTypeThemeConfig;
}